tests/requirements.txt
fabric
fabtools
jinja2
PyYAML
